requirements.txt
streamlit==1.32.0
langchain==0.1.16
langchain-community==0.0.38
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
chromadb==0.5.23
mlflow==2.11.3
python-dotenv==1.0.1
numpy==1.26.4
pypdf==4.1.0
pandas==2.2.1
tqdm==4.66.2
